requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
pyyaml>=5.3.1

# Visualization
matplotlib>=3.2.2
seaborn>=0.11.0
tabulate>=0.8.7

# Data processing
pillow>=7.1.2
tqdm>=4.41.0
scipy>=1.4.1
scikit-learn>=0.24.0

# Utilities
requests>=2.23.0
pandas>=1.1.4
protobuf>=3.19.6

# Optional: Deep learning utilities
tensorboard>=2.4.1

# For development
pytest>=6.0.0
flake8>=3.8.0
black>=21.5b2 